add confirmation when deleting a backup

+++ webui.go
@@ -229,7 +229,7 @@ const backupstemplate = `{{define "BACKUPS"}}{{template "HEADER" .}}
 const backuptemplate = `{{define "BACKUP"}}{{template "HEADER" .}}
 {{with .Backups}}
     {{range .}}
-<div class="submenu">{{if .Files}}<a href="">Open</a><a href="">&#10003; Verify</a>{{end}}<a href="/delete/{{.Name}}/{{.Date}}" class="caution">&#10006; Delete</a></div>
+<div class="submenu">{{if .Files}}<a href="">Open</a><a href="">&#10003; Verify</a>{{end}}<a href="/delete/{{.Name}}/{{.Date}}" onclick="return confirm('Are you sure?')" class="caution">&#10006; Delete</a></div>
 <table class="report">
 <tbody>
 	<tr><th class="rowtitle">ID</th><td class="{{. | status}}" title="{{. | status}}">{{.Date}}</td></tr>
